Put the z3-related routines in a separate class in a separate file
